refactor: estrai domain/catalog.py da cli/
Sposta CatalogResolver da cli/catalog_ops.py a domain/catalog.py.
cli/catalog_ops.py diventa re-export puro.
MCP e test aggiornati a importare da domain/.
Azzera dipendenza MCP→cli/ per catalog_ops.

6 file, 0 logica modificata.
